unloading wifi modules on suspend/resume does more harm than good.
I thought the wifi module was causing my weird wakeup problems before it
turned out to be the USB controller. After installing new firmware on my
router I would be unable to connect to my AP. it seems it never expires
the authentication state of an already authenticated wireless client.

I'm sure that's likely a bug in my router firmware but regardless I
tested just leaving the module inserted and it will save the radio state
and save me from this problem.